Basic tile blocking

+++ src/core/include/ptensor/span3d.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <cassert>
-#include <cstddef>
+#include <cstdint>
 
 #include "span2d.hpp"
 
@@ -11,43 +11,43 @@ class Span3D {
   public:
     Span3D() = default;
 
-    Span3D(T* data, size_t height, size_t width, size_t channels) :
+    Span3D(T* data, int64_t height, int64_t width, int64_t channels) :
         data_(data),
         height_(height),
         width_(width),
         channels_(channels) {}
 
-    T* row(size_t row) {
+    T* row(int64_t row) {
         assert(row < height_);
         return data_ + row * width_ * channels_;
     }
 
-    const T* row(size_t row) const {
+    const T* row(int64_t row) const {
         assert(row < height_);
         return data_ + row * width_ * channels_;
     }
 
-    T* channel(size_t row, size_t col) {
+    T* channel(int64_t row, int64_t col) {
         assert(row < height_);
         assert(col < width_);
         return data_ + row * width_ * channels_ + col * channels_;
     }
 
-    const T* channel(size_t row, size_t col) const {
+    const T* channel(int64_t row, int64_t col) const {
         assert(row < height_);
         assert(col < width_);
         return data_ + row * width_ * channels_ + col * channels_;
     }
 
-    size_t height() const {
+    int64_t height() const {
         return height_;
     }
 
-    size_t width() const {
+    int64_t width() const {
         return width_;
     }
 
-    size_t channels() const {
+    int64_t channels() const {
         return channels_;
     }
 
@@ -61,9 +61,9 @@ class Span3D {
 
   private:
     T* data_ = nullptr;
-    size_t height_ = 0;
-    size_t width_ = 0;
-    size_t channels_ = 0;
+    int64_t height_ = 0;
+    int64_t width_ = 0;
+    int64_t channels_ = 0;
 };
 
 template<typename T>
@@ -71,38 +71,38 @@ class PlanarSpan3D {
   public:
     PlanarSpan3D() = default;
 
-    PlanarSpan3D(T* data, size_t channels, size_t height, size_t width) :
+    PlanarSpan3D(T* data, int64_t channels, int64_t height, int64_t width) :
         data_(data),
         channels_(channels),
         height_(height),
         width_(width) {}
 
-    Span2D<const T> operator[](size_t channel) const {
+    Span2D<const T> operator[](int64_t channel) const {
         assert(channel < channels_);
         return Span2D<const T>(data_ + width_ * height_ * channel, height_, width_);
     }
 
-    Span2D<T> operator[](size_t channel) {
+    Span2D<T> operator[](int64_t channel) {
         assert(channel < channels_);
         return Span2D<T>(data_ + width_ * height_ * channel, height_, width_);
     }
 
-    size_t channels() const {
+    int64_t channels() const {
         return channels_;
     }
 
-    size_t height() const {
+    int64_t height() const {
         return height_;
     }
 
-    size_t width() const {
+    int64_t width() const {
         return width_;
     }
 
   private:
     T* data_ = nullptr;
-    size_t channels_ = 0;
-    size_t height_ = 0;
-    size_t width_ = 0;
+    int64_t channels_ = 0;
+    int64_t height_ = 0;
+    int64_t width_ = 0;
 };
 }  // namespace p10
